improvements from pairing session
